requirements.txt
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.26.0,<2.1.0
scipy>=1.11.4
scikit-learn>=1.1.0
pandas>=1.4.0
matplotlib>=3.5.0
seaborn>=0.11.0
pyyaml>=6.0
tqdm>=4.64.0
pydicom>=2.3.0
opencv-python==4.8.1.78
pillow>=9.0.0
plotly>=5.10.0
captum>=0.5.0
lime>=0.2.0.1
shap>=0.41.0
omegaconf>=2.2.0
hydra-core>=1.2.0
wandb>=0.13.0
tensorboard>=2.9.0
albumentations>=1.2.0
imgaug>=0.4.0
statsmodels>=0.13.0
torchmetrics>=0.7.0
pytorch-lightning>=1.5.0
ipywidgets>=7.6.0
jupyter>=1.0.0